requirements.txt
streamlit>=1.28.0
langgraph>=0.0.20
openai>=1.0.0
tavily-python>=0.3.0
trafilatura>=1.6.0
beautifulsoup4>=4.12.0
requests>=2.31.0